+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v40.x / v41 (In progress)
+- Character stats: Added ability to shuffle special abilities (Pray, Spy, Mirror). Mirror will have an increased chance of success if this mode is activated.
+
 ## v40 (2019-06-09)
 - Enemy stats: Battle palette randomization improved and moved to Palettes flag.
 - Game improvement patches: Ludicrous speed patch now displays new speed names.